python-processor/requirements.txt
Flask==2.3.2
rembg==2.0.50
Pillow==10.0.0
gunicorn==21.2.0